Assignment 3 in CakePHP
It has users, articles, comments, and tags.

An admin user is able to view, edit, create and delete articles,
comments and tags.

A public user is only able to view articles (lists of articles, single
articles as well as view them by tags)  each single article includes
the approved comments for that article.

Comments would not shown for public users (not signed in) unless they
are approved by an admin. An admin would be able to view both approved
/ unapproved comments, and be able to change their status.